+++ downward/suites.py
@@ -9,12 +9,16 @@ def find_domain_file(benchmarks_dir, domain, problem):
     For a given problem filename "<taskname>.<ext>", check the following
     domain filenames: "domain.pddl", "<taskname>-domain.<ext>",
     "domain_<taskname>.<ext>" and "domain-<taskname>.<ext>", where ".<ext>"
-    is optional.
+    is optional. Also check "<xyz>-domain.pddl" where <xyz> are the first
+    three characters of the task file name, to cover the airport domain
+    where task file names are p01-xxx.pddl and domain file names are
+    p01-domain.pddl.
     """
     problem_root, ext = os.path.splitext(problem)
     domain_basenames = [
         "domain.pddl",
         problem_root + "-domain" + ext,
+        problem_root[:3] + "-domain.pddl",  # for airport
         "domain_" + problem,
         "domain-" + problem,
     ]